MH  : move to templates folder
because MH has no decorators
