+++ web/src/engine/websites/Seven8te.ts
@@ -2,7 +2,7 @@ import { Tags } from '../Tags';
 import icon from './Seven8te.webp';
 import { DecoratableMangaScraper } from '../providers/MangaPlugin';
 import * as Common from './decorators/Common';
-import * as MH from './decorators/MH';
+import * as MH from './templates/MH';
 
 @Common.MangaCSS(/^{origin}\/cartoon\/\d+$/, MH.queryMangaTitleFromURI)
 @Common.MangasMultiPageCSS(MH.mangaPath, MH.queryMangas)
