+++ web/src/engine/websites/xianman123.ts
@@ -2,16 +2,11 @@ import { Tags } from '../Tags';
 import icon from './xianman123.webp';
 import { DecoratableMangaScraper } from '../providers/MangaPlugin';
 import * as Common from './decorators/Common';
-import * as MH from './decorators/MH';
+import * as MH from './templates/MH';
 
 const pagesScript = `
-    new Promise((resolve, reject) => {
-        try {
-            const pageLinks = picdata.map(pic => imgDomain != '' ? new URL(pic, imgDomain).href : pic);
-            resolve(pageLinks);
-        } catch (error) {
-            reject(error);
-        }
+    new Promise( resolve  => {
+        resolve( picdata.map(pic => imgDomain != '' ? new URL(pic, imgDomain).href : pic));
     });
  `;
